+++ lib/falcon.rb
@@ -5,6 +5,7 @@
 
 require_relative "falcon/version"
 require_relative "falcon/server"
+require_relative "falcon/composite_server"
 
 # Falcon, running on Rails, requires specific configuration:
 require_relative "falcon/railtie" if defined?(Rails::Railtie)